fix: remove auto-install that didn't work right
Before, if you included the plugin as a script tag, it would
automatically install itself. Because of this behavior, it was
impossible to set global options for the plugin. I'm going to
consider it a beaking change because the behavior was not
documented and broke the behavior that was documented.